casting for signed ops

+++ src/sim.cpp
@@ -205,7 +205,7 @@ int main(int argc, char** argv) {
                         break;
                     case FUN_SLTU: 
                         printf("SLTU!\n");
-                        regData.registers[rd] = (regData.registers[rs] < regData.registers[rt]) ? 1 : 0;
+                        regData.registers[rd] = ((int32_t)regData.registers[rs] < (int32_t)regData.registers[rt]) ? 1 : 0;
                         break;
                     case FUN_SLL: 
                         printf("SLL: %d by %d\n",regData.registers[rt], shamt);
